tested stupid shaders again, added tiny mode to make commands

+++ src/core/rendering/opengl/shaders/textured3D.c
@@ -1,6 +1,6 @@
 const GLchar* shader3D_textured_vert_buffer = "\
 #version 300 es\n\
-layout(location=0) in highp vec3 vertex_position;\
+layout(location=0) in lowp vec3 vertex_position;\
 layout(location=1) in lowp vec2 vertex_uv;\
 layout(location=2) in lowp vec3 vertex_color;\
 uniform highp mat4 camera_matrix;\
